fix: Fix the path parsing problem in load user provided dir test case
Currently when the user provides a base dir to load the env file,
the parsing fails in the windows environment. This is because, the
way path.toString() method used in constructing the path does not
conform to the URI that can be consumed by windows.

This update fixes this error by using URI to construct the full
dirctory path.

Affected File:
 - modified:   src/test/java/io/sysr/springcontext/env/EnvContextLoaderTest.java