feat: Add EnvironmentPostProcessor for loading properties and corresponding unit test
Implemented `SpringContextEnvironmentPostProcessor` to load properties
into the Spring Environment from a custom `EnvContextLoader`.

 The processor adds a `PropertiesPropertySource` named
`springContextDotEnv` to the environment. The implementation ensures
 that the properties are available with the highest precedence by
intergrating `Ordered` funcitionality and setting the order to
`Ordered.HIGHEST_PRECEDENCE`.

Unit Test is added in `SpringContextEnvironmentPostProcessorTest`
to verify the behavior of the `EnvironmentPostProcessor`. The test
verifies that the `postProcessEnvironment` method correctly adds
the `springContextDotEnv` property source.

Affected Files:
 - src/main/java/io/sysr/springcontext/env/SpringContextEnvironmentPostProcessor.java
 - src/test/java/io/sysr/springcontext/env/SpringContextEnvironmentPostProcessorTest.java