fix: Update the variable resolution
Corrected the logic of accessing the loaded variables from the env properties
file. Initially, we were just checking if the value assocciated with the key
is null, not taking into account when it could be empty or blank. In this update,
we ensure that only KEY=VALUE pairs are processed. That is to mean, if the
value associated with the key is empty ot blank, we stop resolving the env
variable, and will be ignored - not included in the final resolved env properties.

In cases where user would like to have the values as empty, for example
`PASSWORD=""` We will not evaluate this, so the resulting properties
will not contain the variable `PASSWORD`. This is an opinionated approach.

Affected files:
 - src/main/java/io/sysr/springcontext/env/EnvContextLoader.java